variable type changed

+++ src/main/java/com/hireon/backend/Model/RequiredSkills.java
@@ -22,7 +22,7 @@ public class RequiredSkills {
     @JoinColumn(name = "skill_id")
     private Skill skill;
 
-    private Integer reqProficiency;
+    private Double reqProficiency;
